Update 7e329abc-0530-7016-7712-687de2c780a8.md
The summary description implied that all documents are always saved.  They aren't.  Hence the change.

I have found that in Office 2010 the save sometimes fails when a Document.Save is used within a VBproject.  It sometimes, but not always, fails when the Find and Replace dialog is open.  It is not clear why it only occurs sometimes and this should be explained.  I do not have the information to clarify this behaviour but Microsoft staff should have it and should place it here.  The seemingly erratic behaviour is a nuisance to programmers which they should be alerted to;  it requires a handler in their code.  (Annoyingly, too a failed save is not trapped by an On Error statement, so the programmer actually NEEDs to know what causes the problem if they are to write suitable error trapping code.)